Postgres Support

+++ bin/index.js
@@ -1,10 +1,21 @@
 const Discord = require("discord.js");
 const fs = require("fs");
 const util = require("plugit-util");
+const pg = require('pg');
+
+
+//Create Clients
+const db = new pg.Client({
+  connectionString: process.env.DATABASE_URL,
+  ssl: true,
+});
 const client = new Discord.Client();
 
+
+//Initialize Variables
 var commands = {}; //Create Dictionary to store Commands
 var events = {};
+
 
 //Help Command
 commands["help"] = {
@@ -24,7 +35,8 @@ commands["help"] = {
   }
 }
 
-var loadCommands = function() {
+
+var loadModules = function() {
     //Load NPM Modules
     var modules = util.modules.getModules();
     for (let file of modules) {
@@ -56,43 +68,62 @@ var loadCommands = function() {
         for (event in module.events) {
           events[event] = module.events[event]
         }
-        if(process.env.debug) {
+        if(process.env.debug == true) {
             console.log("Loaded " + file.slice(0, -3) + " Module");
         }
       }
     }
-    console.log("———— Modules Loaded! ————");
 }
+
 
 client.on('ready', () => {
   console.log(`Logged in as ${client.user.tag}!`);
   console.log(`Connected to ${client.guilds.size} guilds, serving ${client.users.size} users.`);
-  loadCommands();
-  util.modules.logModules(util.modules.getModules());
   client.user.setGame(process.env.playing)
 
+  //Load commands & events
+  loadModules();
+  util.modules.logModules(util.modules.getModules());
+
+  //Handle events
+  if (events['ready']) {
+    events['ready'].main(client, db)
+  }
+
   for (event in events) {
-    client.on(event, (arg1, arg2, arg3) => {
-      events[event].main(client, arg1)
+    if (event == 'message') return
+    client.on(event, (arg1) => {
+      events[event].main(client, db, arg1)
     })
   }
 });
 
+
+//Handle Commands
 client.on('message', msg => {
-  if ( msg.author.bot ) return
+  if ( msg.author.bot ) return //If message is from a bot ignore.
   if ( msg.content.indexOf(process.env.prefix) !== -1 ) {
     var command = msg.content.split(process.env.prefix)[1].split(" ")[0];
     if(command in commands){
-      if(process.env.delete_commands == true){
-         msg.delete();
-      }
-      commands[command].main(client, msg);
+      if(process.env.delete_commands == true) msg.delete()
+      commands[command].main(client, db, msg);
+      return
     }
+    return
   }
 
   if (events["message"]) {
-    events["message"].main(client, msg)
+    events["message"].main(client, db, msg)
   }
 });
+
+
+//Connect to Discord
+console.log("———————— Plugit! ————————");
+
+db.connect()
+  .then(() => console.log(`Successfully Connected To Database`))
+  .catch(e => console.error('Connection error', err.stack))
+
 
 client.login(process.env.token);
